reduce number of files in openvino framework executor
