+++ src/benchmark/frameworks/openvino/openvino_process_factory.py
@@ -1,7 +1,5 @@
-from .openvino_async_process import AsyncOpenVINOProcess
-from .openvino_benchmark_cpp_process import OpenVINOBenchmarkCppProcess
-from .openvino_benchmark_python_process import OpenVINOBenchmarkPythonProcess
-from .openvino_sync_process import SyncOpenVINOProcess
+from .openvino_benchmark_process import OpenVINOBenchmarkPythonProcess, OpenVINOBenchmarkCppProcess
+from .openvino_python_api_process import AsyncOpenVINOProcess, SyncOpenVINOProcess
 
 
 def create_process(test, executor, log, cpp_benchmark_path=None):
