+++ src/benchmark/tests/test_processes.py
@@ -3,15 +3,14 @@ import sys
 
 import pytest
 
-from benchmark.frameworks.processes import ProcessHandler
 from src.benchmark.frameworks.framework_wrapper_registry import FrameworkWrapperRegistry
 from src.benchmark.frameworks.intel_caffe.intel_caffe_process import IntelCaffeProcess
 from src.benchmark.frameworks.known_frameworks import KnownFrameworks
-from src.benchmark.frameworks.openvino.openvino_async_process import AsyncOpenVINOProcess
-from src.benchmark.frameworks.openvino.openvino_benchmark_cpp_process import OpenVINOBenchmarkCppProcess
-from src.benchmark.frameworks.openvino.openvino_benchmark_python_process import OpenVINOBenchmarkPythonProcess
+from src.benchmark.frameworks.openvino.openvino_benchmark_process import OpenVINOBenchmarkPythonProcess, \
+    OpenVINOBenchmarkCppProcess
 from src.benchmark.frameworks.openvino.openvino_process import OpenVINOProcess
-from src.benchmark.frameworks.openvino.openvino_sync_process import SyncOpenVINOProcess
+from src.benchmark.frameworks.openvino.openvino_python_api_process import AsyncOpenVINOProcess, SyncOpenVINOProcess
+from src.benchmark.frameworks.processes import ProcessHandler
 from src.benchmark.frameworks.tensorflow.tensorflow_process import TensorFlowProcess
 from src.benchmark.tests.test_executor import get_host_executor
 
@@ -78,7 +77,7 @@ def test_framework_wrapper(inference_framework, mode, mocker):
 
 def test_get_openvino_benchmark_app_metrics(mocker):
     mocker.patch(
-        'src.benchmark.frameworks.openvino.openvino_benchmark_python_process.OpenVINOBenchmarkPythonProcess._fill_command_line',
+        'src.benchmark.frameworks.openvino.openvino_benchmark_process.OpenVINOBenchmarkPythonProcess._fill_command_line',
         return_value='ls')
     mocker.patch('src.benchmark.executors.HostExecutor.execute_process',
                  return_value=(0, OPENVINO_BENCHMARK_RESULT_RAW.encode('utf-8')))
